requirements.txt
pyrogram
tgcrypto
pyqrcode
pypng
pyzbar
Pillow 
motor
aiofiles
dnspython
